requirements.txt
alembic==1.10.4
anyio==3.6.2
app==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pycountry==22.3.5
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0

pytest~=7.3.1
redis~=4.5.4
alembic==1.10.4
anyio==3.6.2
app==0.0.1
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycountry==22.3.5
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
redis==4.5.4
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
alembic==1.10.4
anyio==3.6.2
app==0.0.1
async-timeout==4.0.2
black==23.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.12.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
pre-commit==3.3.1
psycopg2-binary==2.9.6
pycountry==22.3.5
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.4
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
virtualenv==20.23.0
